下单逻辑，一会儿修改bug

+++ src/models/event_order_new.rs
@@ -1,6 +1,4 @@
 use serde::{Deserialize, Serialize};
-use rust_decimal::Decimal;
-use sqlx::types::BigDecimal;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct EventOrderNew {
@@ -17,19 +15,19 @@ pub struct EventOrderNew {
     pub order_type: Option<String>,  // Adjust this type based on your OrderType enum
 
     #[serde(rename = "price", skip_serializing_if = "Option::is_none")]
-    pub price: Option<BigDecimal>,
+    pub price: Option<f64>,
 
     #[serde(rename = "quantity", skip_serializing_if = "Option::is_none")]
-    pub quantity: Option<BigDecimal>,
+    pub quantity: Option<f64>,
 
     #[serde(rename = "amount", skip_serializing_if = "Option::is_none")]
-    pub amount: Option<BigDecimal>,
+    pub amount: Option<f64>,
 
     #[serde(rename = "max_amount", skip_serializing_if = "Option::is_none")]
-    pub max_amount: Option<BigDecimal>,
+    pub max_amount: Option<f64>,
 
     #[serde(rename = "max_qty", skip_serializing_if = "Option::is_none")]
-    pub max_qty: Option<BigDecimal>,
+    pub max_qty: Option<f64>,
 
     #[serde(rename = "nano_time", skip_serializing_if = "Option::is_none")]
     pub nano_time: Option<i64>,
